Replace InterOpWith with CustomizeOutgoingMessagesOfType
